requirements.txt
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-telegram-bot==20.7